@@ -10,13 +10,14 @@ dependencies:
 dev_dependencies:
   flutter_driver:
     sdk: flutter
+  test: ^1.10.0-nullsafety.1
   integration_test:
     path: ../../../integration_test
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0-nullsafety.1
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.12.13+hotfix.5 <2.0.0"
