@@ -65,8 +65,10 @@ public class Camera {
   private CaptureRequest.Builder captureRequestBuilder;
   private MediaRecorder mediaRecorder;
   private boolean recordingVideo;
+  private File videoRecordingFile;
   private CamcorderProfile recordingProfile;
   private int currentOrientation = ORIENTATION_UNKNOWN;
+  private Context applicationContext;
 
   // Mirrors camera.dart
   public enum ResolutionPreset {
@@ -94,6 +96,7 @@ public class Camera {
     this.flutterTexture = flutterTexture;
     this.dartMessenger = dartMessenger;
     this.cameraManager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
+    this.applicationContext = activity.getApplicationContext();
     orientationEventListener =
         new OrientationEventListener(activity.getApplicationContext()) {
           @Override
@@ -135,7 +138,7 @@ public class Camera {
   }
 
   @SuppressLint("MissingPermission")
-  public void open(@NonNull final Result result) throws CameraAccessException {
+  public void open() throws CameraAccessException {
     pictureImageReader =
         ImageReader.newInstance(
             captureSize.getWidth(), captureSize.getHeight(), ImageFormat.JPEG, 2);
@@ -154,15 +157,13 @@ public class Camera {
             try {
               startPreview();
             } catch (CameraAccessException e) {
-              result.error("CameraAccess", e.getMessage(), null);
+              dartMessenger.sendCameraErrorEvent(e.getMessage());
               close();
               return;
             }
-            Map<String, Object> reply = new HashMap<>();
-            reply.put("textureId", flutterTexture.id());
-            reply.put("previewWidth", previewSize.getWidth());
-            reply.put("previewHeight", previewSize.getHeight());
-            result.success(reply);
+
+            dartMessenger.sendCameraInitializedEvent(
+                previewSize.getWidth(), previewSize.getHeight());
           }
 
           @Override
@@ -174,7 +175,7 @@ public class Camera {
           @Override
           public void onDisconnected(@NonNull CameraDevice cameraDevice) {
             close();
-            dartMessenger.send(DartMessenger.EventType.ERROR, "The camera was disconnected.");
+            dartMessenger.sendCameraErrorEvent("The camera was disconnected.");
           }
 
           @Override
@@ -200,7 +201,7 @@ public class Camera {
               default:
                 errorDescription = "Unknown camera error";
             }
-            dartMessenger.send(DartMessenger.EventType.ERROR, errorDescription);
+            dartMessenger.sendCameraErrorEvent(errorDescription);
           }
         },
         null);
@@ -218,12 +219,13 @@ public class Camera {
     return flutterTexture;
   }
 
-  public void takePicture(String filePath, @NonNull final Result result) {
-    final File file = new File(filePath);
-
-    if (file.exists()) {
-      result.error(
-          "fileExists", "File at path '" + filePath + "' already exists. Cannot overwrite.", null);
+  public void takePicture(@NonNull final Result result) {
+    final File outputDir = applicationContext.getCacheDir();
+    final File file;
+    try {
+      file = File.createTempFile("CAP", ".jpg", outputDir);
+    } catch (IOException | SecurityException e) {
+      result.error("cannotCreateFile", e.getMessage(), null);
       return;
     }
 
@@ -232,7 +234,7 @@ public class Camera {
           try (Image image = reader.acquireLatestImage()) {
             ByteBuffer buffer = image.getPlanes()[0].getBuffer();
             writeToFile(buffer, file);
-            result.success(null);
+            result.success(file.getAbsolutePath());
           } catch (IOException e) {
             result.error("IOError", "Failed saving image", null);
           }
@@ -308,8 +310,7 @@ public class Camera {
           public void onConfigured(@NonNull CameraCaptureSession session) {
             try {
               if (cameraDevice == null) {
-                dartMessenger.send(
-                    DartMessenger.EventType.ERROR, "The camera was closed during configuration.");
+                dartMessenger.sendCameraErrorEvent("The camera was closed during configuration.");
                 return;
               }
               cameraCaptureSession = session;
@@ -320,14 +321,13 @@ public class Camera {
                 onSuccessCallback.run();
               }
             } catch (CameraAccessException | IllegalStateException | IllegalArgumentException e) {
-              dartMessenger.send(DartMessenger.EventType.ERROR, e.getMessage());
+              dartMessenger.sendCameraErrorEvent(e.getMessage());
             }
           }
 
           @Override
           public void onConfigureFailed(@NonNull CameraCaptureSession cameraCaptureSession) {
-            dartMessenger.send(
-                DartMessenger.EventType.ERROR, "Failed to configure camera session.");
+            dartMessenger.sendCameraErrorEvent("Failed to configure camera session.");
           }
         };
 
@@ -369,18 +369,24 @@ public class Camera {
     cameraDevice.createCaptureSession(surfaces, callback, null);
   }
 
-  public void startVideoRecording(String filePath, Result result) {
-    if (new File(filePath).exists()) {
-      result.error("fileExists", "File at path '" + filePath + "' already exists.", null);
+  public void startVideoRecording(Result result) {
+    final File outputDir = applicationContext.getCacheDir();
+    try {
+      videoRecordingFile = File.createTempFile("REC", ".mp4", outputDir);
+    } catch (IOException | SecurityException e) {
+      result.error("cannotCreateFile", e.getMessage(), null);
       return;
     }
+
     try {
-      prepareMediaRecorder(filePath);
+      prepareMediaRecorder(videoRecordingFile.getAbsolutePath());
       recordingVideo = true;
       createCaptureSession(
           CameraDevice.TEMPLATE_RECORD, () -> mediaRecorder.start(), mediaRecorder.getSurface());
       result.success(null);
     } catch (CameraAccessException | IOException e) {
+      recordingVideo = false;
+      videoRecordingFile = null;
       result.error("videoRecordingFailed", e.getMessage(), null);
     }
   }
@@ -396,7 +402,8 @@ public class Camera {
       mediaRecorder.stop();
       mediaRecorder.reset();
       startPreview();
-      result.success(null);
+      result.success(videoRecordingFile.getAbsolutePath());
+      videoRecordingFile = null;
     } catch (CameraAccessException | IllegalStateException e) {
       result.error("videoRecordingFailed", e.getMessage(), null);
     }
