@@ -2,6 +2,12 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// TODO(amirh): Remove this once flutter_driver supports null safety.
+// https://github.com/flutter/flutter/issues/71379
+// @dart = 2.9
+import 'dart:async';
+
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -22,7 +28,7 @@ void main() {
     testWidgets('can be initialized', (WidgetTester tester) async {
       await _controller.initialize();
 
-      expect(_controller.value.initialized, true);
+      expect(_controller.value.isInitialized, true);
       expect(_controller.value.position, const Duration(seconds: 0));
       expect(_controller.value.isPlaying, false);
       expect(_controller.value.duration,
@@ -30,9 +36,57 @@ void main() {
     });
 
     testWidgets(
+      'reports buffering status',
+      (WidgetTester tester) async {
+        VideoPlayerController networkController = VideoPlayerController.network(
+          'https://flutter.github.io/assets-for-api-docs/assets/videos/bee.mp4',
+        );
+        await networkController.initialize();
+        // Mute to allow playing without DOM interaction on Web.
+        // See https://developers.google.com/web/updates/2017/09/autoplay-policy-changes
+        await networkController.setVolume(0);
+        final Completer<void> started = Completer();
+        final Completer<void> ended = Completer();
+        bool startedBuffering = false;
+        bool endedBuffering = false;
+        networkController.addListener(() {
+          if (networkController.value.isBuffering && !startedBuffering) {
+            startedBuffering = true;
+            started.complete();
+          }
+          if (startedBuffering &&
+              !networkController.value.isBuffering &&
+              !endedBuffering) {
+            endedBuffering = true;
+            ended.complete();
+          }
+        });
+
+        await networkController.play();
+        await networkController.seekTo(const Duration(seconds: 5));
+        await tester.pumpAndSettle(_playDuration);
+        await networkController.pause();
+
+        expect(networkController.value.isPlaying, false);
+        expect(networkController.value.position,
+            (Duration position) => position > const Duration(seconds: 0));
+
+        await started;
+        expect(startedBuffering, true);
+
+        await ended;
+        expect(endedBuffering, true);
+      },
+      skip: !(kIsWeb || defaultTargetPlatform == TargetPlatform.android),
+    );
+
+    testWidgets(
       'can be played',
       (WidgetTester tester) async {
         await _controller.initialize();
+        // Mute to allow playing without DOM interaction on Web.
+        // See https://developers.google.com/web/updates/2017/09/autoplay-policy-changes
+        await _controller.setVolume(0);
 
         await _controller.play();
         await tester.pumpAndSettle(_playDuration);
@@ -58,6 +112,9 @@ void main() {
       'can be paused',
       (WidgetTester tester) async {
         await _controller.initialize();
+        // Mute to allow playing without DOM interaction on Web.
+        // See https://developers.google.com/web/updates/2017/09/autoplay-policy-changes
+        await _controller.setVolume(0);
 
         // Play for a second, then pause, and then wait a second.
         await _controller.play();
@@ -104,6 +161,6 @@ void main() {
 
       await tester.pumpAndSettle();
       expect(_controller.value.isPlaying, true);
-    });
+    }, skip: kIsWeb); // Web does not support local assets.
   });
 }
